+++ CHANGELOG.md
@@ -2,15 +2,30 @@
 
 # UNRELEASED
 
-* AADGroupOwnerConsentSettings
-  * Initial release
-    Implements [#4112](https://github.com/microsoft/Microsoft365DSC/issues/4112)
+
 * AADAdministrativeUnit
   * Fix the Update logic flow to get around a bug in Microsoft.Graph 2.11.1.
+* AADAuthenticationMethodPolicyX509
+  * Added support for the  property for include targets
 * AADConditionalAccessPolicy
   * Added support for application filters in the conditions.
 * AADConditionalAccessPolicy
-  * FIXES [[#3885](https://github.com/microsoft/Microsoft365DSC/issues/3885)]
+  * Implement Fix #3885. Manage Exclude Application.
+    FIXES [[#3885](https://github.com/microsoft/Microsoft365DSC/issues/3885)]
+* AADGroupOwnerConsentSettings
+  * Initial release
+    Implements [#4112](https://github.com/microsoft/Microsoft365DSC/issues/4112)
+* EXOHostedContentFilterPolicy
+  * Fix issue on parameters AllowedSenders, AllowedSenderDomains, BlockedSenders, 
+    BlockSenderDomains if desired state is empty but current state is not empty.
+    FIXES[#4124](https://github.com/microsoft/Microsoft365DSC/issues/4124)
+* IntuneDeviceConfigurationPolicyMacOS
+  * Fix workaround added on PR #4099 in order to be able to use this resource
+    for deployments
+    FIXES [#4105](https://github.com/microsoft/Microsoft365DSC/issues/4105)
+* SCDLPComplianceRule
+  * Fix type of AccessScope
+    FIXES [#3463](https://github.com/microsoft/Microsoft365DSC/issues/3463)
 
 # 1.24.103.1